feat.lang novas alterações

+++ Questao07/Aluno.py
@@ -1,7 +1,7 @@
 nome = input("Escreva o nome do aluno: ")
 disciplina = input("Escreva o nome da disciplina: ")
-nota1 = int(input("Digite a nota 1: "))
-nota2 = int(input("Digite a nota 2: "))
+nota1 = float(input("Digite a nota 1: "))
+nota2 = float(input("Digite a nota 2: "))
 m = (nota1 + nota2) /2
 
 print(f'A média do aluno é {m}')
